requirements.txt
altgraph==0.17
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.4.0
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.3
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.1
Django==3.2
django-adminlte3==0.1.6
django-allauth==0.45.0
django-ckeditor==6.1.0
django-cleanup==5.2.0
django-crispy-forms==1.12.0
django-js-asset==1.2.2
django-material==1.9.0
django-material-admin==1.8.5
django-suit==0.2.25
djangorestframework==3.12.4
filelock==3.0.12
future==0.18.2
idna==2.10
importlib-metadata==4.6.1
macholib==1.14
MouseInfo==0.1.3
oauthlib==3.1.1
pefile==2019.4.18
Pillow==8.2.0
psycopg2==2.9.1
PyAutoGUI==0.9.52
pycparser==2.20
PyGetWindow==0.0.9
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
PyJWT==2.1.0
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.26
python3-openid==3.2.0
PyTweening==1.0.3
pytz==2021.1
pywin32-ctypes==0.2.0
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.3
virtualenv==20.4.3
zipp==3.5.0
